Move to bundler 2.1
I was able to get the package building in Habitat with this change but
haven't tested it beyond that.

This fixes the Habitat build because, as far as I can tell, the
breakage looks something like this:

1) Bundler 2 has a feature where, when it detects a Gemfile bundled
   with an older Gemfile, it tries to switch to that major version.

2) In the case of core/bundler this meant that when you ran bundle
   inside this repository, it would report its version as 1.16.6 which
   is the version that shipped with the version of Ruby core/bundler
   is using.

3) The scaffolding assumes that it can always find a gem for the
   version reported by `bundle --version` in the /hab/pkg path for
   core/bundler.  However, because of the version-switching feature,
   this is not the case.

   https://github.com/habitat-sh/core-plans/blob/master/scaffolding-ruby/lib/scaffolding.sh#L350-L352

Upgrading to 2.1 in Gemfile means that core/bundler doesn't switch to
a 1.x version and thus the version reported by `bundle --version` is
indeed available for the scaffolding to vendor.

Signed-off-by: Steven Danna <steve@chef.io>